Better NaN handling for AMDGCN fmed3.
All operands are checked for NaN now. The checks were moved before the
canonicalization to provide a better mapping from fclamp. Changed the
behaviour of fmed3(x,y,NaN) to return max(x,y) instead of min(x,y) in
light of this. Updated tests as a result and added some new cases to
cover the fix.